web-audio/audio-context.html

<!doctype html>
<html class="a">
 <head>
  <title>AudioContext interface</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/resources/lodash.js"></script>
  <script src="/resources/vendor-prefixes.js"></script>
  <script src="js/helpers.js"></script>
 </head>
 <body class="a">
  <div id="log"></div>
  <script>

var attributes = [
  "destination",
  "sampleRate",
  "currentTime",
  "listener",
  "activeSourceCount"
];

var methods = [
  "createBuffer",
  "decodeAudioData",
  "createBufferSource",
  "createMediaElementSource",
  "createMediaStreamSource",
  "createScriptProcessor",
  "createAnalyser",
  "createGain",
  "createDelay",
  "createBiquadFilter",
  "createWaveShaper",
  "createPanner",
  "createConvolver",
  "createChannelSplitter",
  "createChannelMerger",
  "createDynamicsCompressor",
  "createOscillator",
  "createWaveTable"
];

var context;
setup(function() {
  context = new AudioContext();
});

_(attributes).forEach(function(attr) {
  test(function() {
    assert_defined(context, "context is defined")
    assert_has_property(context, attr, "should be an attribute");
    assert_readonly(context, attr, "should be read-only");
  }, attr,
  { help: "http://www.w3.org/TR/webaudio/#AudioContext-section" }
  );
});

_(methods).forEach(function(method) {
  test(function() {
    assert_defined(context, "context is defined")
    assert_is_method(context, method, "should be a method");
  }, method,
  { help: "http://www.w3.org/TR/webaudio/#AudioContext-section" }
  );
});

  </script>
 </body>
</html>
